Rename FramebufferVulkan.cpp/h to FramebufferManagerVulkan.cpp/h
