+++ GPU/Vulkan/FramebufferManagerVulkan.cpp
@@ -43,7 +43,7 @@
 #include "GPU/Common/TextureDecoder.h"
 #include "GPU/Common/FramebufferManagerCommon.h"
 #include "GPU/Debugger/Stepping.h"
-#include "GPU/Vulkan/FramebufferVulkan.h"
+#include "GPU/Vulkan/FramebufferManagerVulkan.h"
 #include "GPU/Vulkan/DrawEngineVulkan.h"
 #include "GPU/Vulkan/TextureCacheVulkan.h"
 #include "GPU/Vulkan/ShaderManagerVulkan.h"
